Restructuring of packages

+++ src/main/java/com/ncedu/testing/web/courses/CourseController.java
@@ -7,7 +7,6 @@ import org.springframework.ui.Model;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestMethod;
-import org.springframework.web.bind.annotation.ResponseBody;
 
 import java.sql.SQLException;
 
@@ -15,7 +14,6 @@ import java.sql.SQLException;
  * Created by ilya on 25.03.17.
  */
 
-//@RequestMapping("student/courses")
 @Controller
 public class CourseController {
 
@@ -35,7 +33,7 @@ public class CourseController {
     @RequestMapping(value = "student/courses/{course_id}", method = RequestMethod.GET)
     public String showCourse(@PathVariable long course_id, Model model) {
         try {
-            model.addAttribute("course", courseService.get((long) course_id));
+            model.addAttribute("course", courseService.get(course_id));
         } catch (SQLException e) {
             e.printStackTrace();
         }
